Move CorsHeadersMiddleware to more specific Http\Auth namespace
