+++ src/Http/Auth/CorsHeadersMiddleware.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace CultuurNet\UDB3\Http\Middleware;
+namespace CultuurNet\UDB3\Http\Auth;
 
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
