+++ tests/Http/Auth/CorsHeadersMiddlewareTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace CultuurNet\UDB3\Http\Middleware;
+namespace CultuurNet\UDB3\Http\Auth;
 
 use CultuurNet\UDB3\Http\Request\Psr7RequestBuilder;
 use CultuurNet\UDB3\Http\Response\NoContentResponse;
